moved userTrip loading function to App.jsx
